requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver
beautifulsoup4
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
redis==5.0.1
adb-shell==0.4.4
websocket-client>=1.6.0
requests>=2.31.0
aiohttp>=3.9.0
PySocks>=1.7.1
APScheduler==3.10.4
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
celery>=5.3.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.0.0
APScheduler==3.10.4
# Sentence Transformers
# sentence-transformers==2.2.2
# torch==2.0.1
# transformers==4.30.0
# huggingface_hub==0.15.1 #    Uninstalling huggingface-hub-0.36.0: